requirements.txt
numpy==2.2.4
torch==2.7.0
tqdm==4.67.1
scikit-learn==1.6.1
dill==0.3.8
rdkit-pypi==2022.9.5
